Uint112 upgraded to uint256, various gas fixes

+++ contracts/v076/OVLTokenTypes.sol
@@ -2,8 +2,8 @@ pragma solidity ^0.7.6;
 // DELTA-BUG-BOUNTY
 
 struct VestingTransaction {
-    uint112 amount;
-    uint112 fullVestingTimestamp;
+    uint256 amount;
+    uint256 fullVestingTimestamp;
 }
 
 struct WalletTotals {
@@ -15,10 +15,10 @@ struct WalletTotals {
 struct UserInformation {
     // TODO move this into better structs? 
     // This is going to be read from only [0]
-    uint8 mostMatureTxIndex;
-    uint8 lastInTxIndex;
-    uint112 maturedBalance;
-    uint112 maxBalance;
+    uint256 mostMatureTxIndex;
+    uint256 lastInTxIndex;
+    uint256 maturedBalance;
+    uint256 maxBalance;
     bool fullSenderWhitelisted;
     // Note that recieving immature balances doesnt mean they recieve them fully vested just that senders can do it
     bool immatureRecieverWhiteslited;
@@ -26,9 +26,9 @@ struct UserInformation {
 }
 
 struct VestingTransactionDetailed {
-    uint112 amount;
-    uint112 fullVestingTimestamp;
+    uint256 amount;
+    uint256 fullVestingTimestamp;
     // uint256 percentVestedE4;
-    uint112 mature;
-    uint112 immature;
+    uint256 mature;
+    uint256 immature;
 }
